src/agents/financial_agent/requirements.txt
# Financial Agent Dependencies

# Web Framework
fastapi>=0.118.0
uvicorn[standard]>=0.27.0.post1
sse-starlette>=2.1.2

# OpenBB AI Integration
openbb-ai>=1.7.5

# LLM
openai>=1.0.0

# Environment Management
python-dotenv>=1.0.1

# HTTP Client
httpx>=0.28.1

# Async Utilities
asyncstdlib>=3.13.1

# Testing
pytest>=8.3.1
pytest-asyncio>=0.23.8

# Data Processing (for potential enhancements)
pandas>=2.0.0
numpy>=1.24.0

# Type Checking (development)
mypy>=1.15.0
